+++ RELEASE-NOTES.md
@@ -36,11 +36,14 @@ END TEMPLATE-->
 ### Breaking changes
 
 * Undid `*.yaml` prototype loading change from previous version.
+* `IConsoleHost`'s `RegisteredCommands` field has been renamed to `AvailableCommands`.
 * Several light related cvars have been renamed. E.g., "display.softshadows" is now "light.softshadows". 
 * The "display.lightmapdivider" integer cvar has been replaced with a float multiplier named "light.resolution_scale".
 
+
 ### New features
 
+* Command definitions have a new bool that restricts them to only be executable by the server or in single player mode. Several "server only" commands have been moved to to shared code and now use this option.
 * The FOV color is now configurable via the "render.fov_color" cvar
 
 ### Bugfixes